code/izutov-evgenii/Infrastructure/src/main/java/ru/unn/agile/DemandElasticity/infrastructure/LoggerToTxt.java
package ru.unn.agile.DemandElasticity.infrastructure;

import java.io.BufferedReader;
import java.io.BufferedWriter;
import java.io.FileReader;
import java.io.FileWriter;
import java.text.SimpleDateFormat;
import java.util.ArrayList;
import java.util.Calendar;
import java.util.List;
import java.util.Locale;

import ru.unn.agile.DemandElasticity.viewmodel.ILogger;

public final class LoggerToTxt implements ILogger {
    private static final String DATE_FORMAT_NOW = "yyyy-MM-dd HH:mm:ss";
    private final BufferedWriter writer;
    private final String fileName;

    public LoggerToTxt(final String fileName) {
        if (fileName == null) {
            throw new IllegalArgumentException("FileName can not be null");
        }

        if (fileName.isEmpty()) {
            throw new IllegalArgumentException("FileName can not be empty");
        }

        this.fileName = fileName;

        BufferedWriter logWriter = null;
        try {
            logWriter = new BufferedWriter(new FileWriter(fileName));
        } catch (Exception e) {
            e.printStackTrace();
        }
        writer = logWriter;
    }

    @Override
    public void logMessage(final String s) {
        try {
            writer.write(curTime() + " > " + s);
            writer.newLine();
            writer.flush();
        } catch (Exception e) {
            System.out.println(e.getMessage());
        }
    }

    @Override
    public List<String> getFullLog() {
        BufferedReader reader;
        ArrayList<String> logMessages = new ArrayList<String>();
        try {
            reader = new BufferedReader(new FileReader(fileName));
            String line = reader.readLine();

            while (line != null) {
                logMessages.add(line);
                line = reader.readLine();
            }
        } catch (Exception e) {
            System.out.println(e.getMessage());
        }

        return logMessages;
    }

    private static String curTime() {
        Calendar cal = Calendar.getInstance();
        SimpleDateFormat sdf = new SimpleDateFormat(DATE_FORMAT_NOW, Locale.ENGLISH);
        return sdf.format(cal.getTime());
    }
}
